fix(setup): extract BALANCED_ANCHORS to client-safe constants module
provider-models.ts imports getSetting (server-only) which transitively
pulls in db/postgres. Importing BALANCED_ANCHORS from that module in a
"use client" page triggered a Turbopack build error (can't resolve 'fs').

Move BALANCED_ANCHORS to provider-model-constants.ts (no server imports)
and re-export it from provider-models.ts for backwards compatibility.
The setup/provider page now imports directly from the constants module.
